security/ethical_guardrails/requirements.txt
transformers>=4.30.0
torch>=2.0.0
spacy>=3.5.0
scikit-learn>=1.2.0
numpy>=1.24.0
pandas>=2.0.0
boto3>=1.28.0
google-cloud-storage>=2.11.0
python-dateutil>=2.8.2
pyyaml>=6.0.0
pytest>=7.3.0
black>=23.3.0
